+++ front/src/components/ScheduleTimeList/index.tsx
@@ -1,28 +1,67 @@
-import { useState } from 'react';
+import { useContext, useState } from 'react';
 
 import Conditional from '@components/Conditional';
+import useSnackbar from '@hooks/useSnackbar';
+import { UserStateContext } from '@context/UserProvider';
 import { getHourMinutes } from '@utils/date';
 import type { TimeSchedule } from '@typings/domain';
+import api from '@api/index';
 import * as S from './styles';
 
-interface ReservationTimeListProps {
+interface ScheduleTimeListProps {
+  date: string;
   daySchedule: TimeSchedule[];
   onClickTime: (dateTime: string) => void;
   onSelectAll: (isSelected: boolean) => void;
-  onUpdateSchedule: () => Promise<void>;
+  onUpdateSchedule: (selectedTimes: string[]) => void;
 }
 
-const ReservationTimeList = ({
+const ScheduleTimeList = ({
+  date,
   daySchedule,
   onClickTime,
   onSelectAll,
   onUpdateSchedule,
-}: ReservationTimeListProps) => {
+}: ScheduleTimeListProps) => {
+  const { userData } = useContext(UserStateContext);
+  const showSnackbar = useSnackbar();
   const [isSelectedAll, setIsSelectedAll] = useState(false);
+
+  const getSelectedTimes = () => {
+    return daySchedule.reduce((newArray, { isSelected, dateTime }) => {
+      if (isSelected) {
+        newArray.push(dateTime);
+      }
+      return newArray;
+    }, [] as string[]);
+  };
 
   const handleSelectAll = () => {
     setIsSelectedAll((prev) => !prev);
     onSelectAll(isSelectedAll);
+  };
+
+  const handleUpdateDaySchedule = async () => {
+    const selectedTimes = getSelectedTimes();
+    try {
+      await api.put(
+        `/api/v2/coaches/me/schedules`,
+        {
+          date,
+          schedules: selectedTimes,
+        },
+        {
+          headers: {
+            Authorization: `Bearer ${userData?.token}`,
+          },
+        }
+      );
+      onUpdateSchedule(selectedTimes);
+      showSnackbar({ message: '확정되었습니다. ✅' });
+    } catch (error) {
+      alert(error);
+      console.log(error);
+    }
   };
 
   return (
@@ -50,11 +89,11 @@ const ReservationTimeList = ({
           <S.CheckButton onClick={handleSelectAll}>
             {isSelectedAll ? '전체 해제' : '전체 선택'}
           </S.CheckButton>
-          <S.ConfirmButton onClick={onUpdateSchedule}>확인</S.ConfirmButton>
+          <S.ConfirmButton onClick={handleUpdateDaySchedule}>확인</S.ConfirmButton>
         </S.ButtonContainer>
       </Conditional>
     </S.TimeListContainer>
   );
 };
 
-export default ReservationTimeList;
+export default ScheduleTimeList;
